updated shared function with a non-overwritable shared library
package.json get's a variable which stores the shared library so that we can track if there is already a shared lib present

+++ generators/createBanner/plain.js
@@ -12,34 +12,50 @@ module.exports = class extends Generator {
   async action() {
     const [width, height] = this.options.size.split('x');
 
-    this.fs.extendJSON(
-      this.destinationPath('package.json'),
-      this.fs.readJSON(this.templatePath('plain/extendPackageJson.json')),
-    );
+    if(!this.options.folderSharedExists) {
+      const jsonPackage = deepmerge(this.fs.readJSON(this.templatePath('plain/extendPackageJson.json')), {
+        shared: this.options.sharedFolder,
+      });
+
+      this.fs.extendJSON(
+        this.destinationPath('package.json'),
+        jsonPackage,
+      );
+    }
+
+
+
+
 
     this.fs.copyTpl(
       this.templatePath('plain/banner'),
       this.destinationPath(path.join(this.options.outputPath)),
       {
-        shared: this.options.isShared,
+        shared: this.options.sharedFolder,
         banner_width: width,
         banner_height: height,
+        bannerPath: "'../../../"+this.options.sharedFolder+"/script/Banner'",
+        animationPath: "'../../../"+this.options.sharedFolder+"/script/Animation'",
       },
     );
 
     if(this.options.isShared) {
-      this.fs.copy(
-        this.templatePath('plain/shared/img'),
-        this.destinationPath(path.join(this.options.outputPath), '../../shared/img'),
-      );
+      if(!this.options.folderSharedExists) {
+        this.fs.copy(
+          this.templatePath('plain/shared/img'),
+          this.destinationPath(path.join(this.options.outputPath), '../../' + this.options.sharedFolder + '/img'),
+        );
 
-      this.fs.copyTpl(
-        this.templatePath('plain/shared/script'),
-        this.destinationPath(path.join(this.options.outputPath, '../../shared/script')),
-        {
-          shared: this.options.isShared,
-        },
-      );
+        this.fs.copyTpl(
+          this.templatePath('plain/shared/script'),
+          this.destinationPath(path.join(this.options.outputPath, '../../' + this.options.sharedFolder + '/script')),
+          {
+            shared: this.options.isShared,
+            bannerPath: "'../../../" + this.options.sharedFolder + "/script/Banner'",
+            animationPath: "'../../../" + this.options.sharedFolder + "/script/Animation'",
+          },
+        );
+      }
     } else {
       this.fs.copy(
         this.templatePath('plain/shared/img'),
@@ -50,7 +66,7 @@ module.exports = class extends Generator {
         this.templatePath('plain/shared/script'),
         this.destinationPath(path.join(this.options.outputPath, 'script')),
         {
-          shared: this.options.isShared,
+          shared: this.options.isShared
         },
       );
     }
@@ -58,7 +74,7 @@ module.exports = class extends Generator {
 
     const json = deepmerge(this.fs.readJSON(this.templatePath('plain/banner/.richmediarc')), {
       content: {
-        logo: this.options.isShared ? "../../shared/img/logo.svg" : "./img/logo.svg",
+        logo: this.options.isShared ? "../../"+this.options.sharedFolder+"/img/logo.svg" : "./img/logo.svg",
       },
       settings: {
         type: this.options.isShared ? "plain-shared" : "plain",
